Add files via upload
The code plots the two subsets of data (CNTL and Treatment animals csv files)  in that it divides the data into two subsets and based on the values in the 'ROI_location' column, using the isin method to select specific rows with specific values in this column. Then using seaborn.scatterplot, and sets the x-axis and y-axis labels to 'Frame' and 'Dis_cm_sum', respectively. Also perform a t-test and F-test on the Velocity data.